# server/swagger.yaml
openapi: 3.0.0
info:
  title: BitcoYemen API
  contact: {}
  version: "V1.2"
servers:
  - url: https://bitco-ye.onrender.com/api/v1
    variables: {}
  - url: https://bitco-ye.herokuapp.com/api/v1
    variables: {}
  - url: http://localhost:5000/api/v1
    variables: {}
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Auth
      summary: Verify Email
      operationId: VerifyEmail
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VerifyEmailRequest"
                - example:
                    email: test@test.com
                    verificationToken: 055ee337407a272f3b393a740013d0ea22338a8da0a7e05293eb60dda836c4007ddd236f00ffadea
            example:
              email: test@test.com
              verificationToken: 055ee337407a272f3b393a740013d0ea22338a8da0a7e05293eb60dda836c4007ddd236f00ffadea
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Forgot Password
      operationId: ForgotPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ForgotPasswordRequest"
                - example:
                    email: test@test.com
            example:
              email: test@test.com
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset Password
      operationId: ResetPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ResetPasswordRequest"
                - example:
                    token: 4078b45e72746d26a252c834e2cbeadea57f8cd5a37193d964e4405654dfd8c2368b10e566c033966ce48e330db5d0dac82c3efc36c762d7ad59b50a088f1175d965630f8b4c
                    email: test@test.com
                    password: "1234567"
            example:
              token: 4078b45e72746d26a252c834e2cbeadea57f8cd5a37193d964e4405654dfd8c2368b10e566c033966ce48e330db5d0dac82c3efc36c762d7ad59b50a088f1175d965630f8b4c
              email: test@test.com
              password: "1234567"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
...
  /currency/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The currency ID
    post:
      tags:
        - Currency
      summary: Update Currency
      operationId: UpdateCurrency
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
                - name
                - code
                - isActive
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example: Bitco
                code:
                  type: string
                  example: BCO
                isActive:
                  type: boolean
                  example: false
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /currency/activate/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The currency ID
    post:
      tags:
        - Currency
      summary: Activate Currency
      operationId: ActivateCurrency
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ActivateCurrencyRequest"
                - example:
                    isActive: false
            example:
              isActive: false
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /direction:
    post:
      tags:
        - Exhange Direction
      summary: Create Exchange Direction
      operationId: CreateExchangeDirection
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateExchangeDirectionRequest"
                - example:
                    from: 62e4c4a5bb77243116e4c386
                    to: 62e4c4d1bb77243116e4c389
                    exchangeRate:
                      send: "1"
                      receive: "1.05"
                    minExchange: "100"
                    maxExchange: "500"
                    description: " "
                    isActive: true
            example:
              from: 62e4c4a5bb77243116e4c386
              to: 62e4c4d1bb77243116e4c389
              exchangeRate:
                send: "1"
                receive: "1.05"
              minExchange: "100"
              maxExchange: "500"
              description: " "
              isActive: true
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Exhange Direction
      summary: Exchange Direction
      operationId: ExchangeDirection
      parameters:
        - name: from
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: 62e4c4a5bb77243116e4c386
        - name: to
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: 62ef5af171bf7efa0ec50317
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
...
  /product:
    post:
      tags:
        - Product
      summary: Create Product
      operationId: CreateProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
                - image
                - name
                - price
                - details
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example: product name
                price:
                  type: integer
                  format: int32
                  example: 50
                details:
                  type: string
                  example: "asd fdsdf "
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Product
      summary: All Products
      operationId: AllProducts
      parameters:
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /product/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The product ID
    post:
      tags:
        - Product
      summary: Update Product
      operationId: UpdateProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
                - name
                - price
                - details
              type: object
              properties:
                image:
                  type: string
                  format: binary
                name:
                  type: string
                  example: product name Up
                price:
                  type: integer
                  format: int32
                  example: 60
                details:
                  type: string
                  example: asd fdsdf 123
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Product
      summary: Product by Id
      operationId: ProductbyId
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    VerifyEmailRequest:
      title: VerifyEmailRequest
      required:
        - email
        - verificationToken
      type: object
      properties:
        email:
          type: string
        verificationToken:
          type: string
      example:
        email: test@test.com
        verificationToken: 055ee337407a272f3b393a740013d0ea22338a8da0a7e05293eb60dda836c4007ddd236f00ffadea
    ForgotPasswordRequest:
      title: ForgotPasswordRequest
      required:
        - email
      type: object
      properties:
        email:
          type: string
      example:
        email: test@test.com
    ResetPasswordRequest:
      title: ResetPasswordRequest
      required:
        - token
        - email
        - password
      type: object
      properties:
        token:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        token: 4078b45e72746d26a252c834e2cbeadea57f8cd5a37193d964e4405654dfd8c2368b10e566c033966ce48e330db5d0dac82c3efc36c762d7ad59b50a088f1175d965630f8b4c
        email: test@test.com
        password: "1234567"
...
    ActivateCurrencyRequest:
      title: ActivateCurrencyRequest
      required:
        - isActive
      type: object
      properties:
        isActive:
          type: boolean
      example:
        isActive: false
    CreateExchangeDirectionRequest:
      title: CreateExchangeDirectionRequest
      required:
        - from
        - to
        - exchangeRate
        - minExchange
        - maxExchange
        - description
        - isActive
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        exchangeRate:
          allOf:
            - $ref: "#/components/schemas/ExchangeRate"
            - {}
        minExchange:
          type: string
        maxExchange:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      example:
        from: 62e4c4a5bb77243116e4c386
        to: 62e4c4d1bb77243116e4c389
        exchangeRate:
          send: "1"
          receive: "1.05"
        minExchange: "100"
        maxExchange: "500"
        description: " "
        isActive: true
    ExchangeRate:
      title: ExchangeRate
      required:
        - send
        - receive
      type: object
      properties:
        send:
          type: string
        receive:
          type: string
      example:
        send: "1"
        receive: "1.05"
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...